requirements.txt
sounddevice
numpy
scipy
openai-whisper
torch
fastapi 
uvicorn 
python-multipart